requirements.txt
gradio
smoalgents
Pillow
streamlit